requirements.txt
langchain
langgraph
sentence-transformers
chromadb
faiss-cpu
tiktoken
transformers
openai
tree_sitter
peft
accelerate
pyyaml
typer
rich
whisper
streamlit
pyqt5
pyinstaller
requests 
